requirements.txt
# Web
Flask>=2.0.0
gunicorn>=21.2.0

# Core
requests>=2.26.0
Pillow>=9.0.0
numpy>=1.24.0
python-dotenv>=1.0.0

# API
aiohttp>=3.8.1
aiofiles>=0.8.0
psutil>=5.8.0

# MCP (optional for local development)
fastmcp==0.4.1
mcp==1.2.1
brotlipy>=0.7.0

# Optional for image processing
opencv-python

# Comment out these to avoid installation issues on Render
# GFPGAN
# realesrgan
# torch==1.13.1
# torchvision==0.14.1